fix data type of coupon and tax rate value

+++ src/main/java/com/invoiced/entity/Coupon.java
@@ -25,7 +25,7 @@ public final class Coupon extends AbstractEntity<Coupon> {
 
     @JsonInclude(JsonInclude.Include.NON_EMPTY)
     @JsonProperty("value")
-    public Long value;
+    public Double value;
 
     @JsonInclude(JsonInclude.Include.NON_EMPTY)
     @JsonProperty("is_percent")
